Move find_color_card to a separate module
